use same type for hash and type index

+++ include/static_type_info/type_index.h
@@ -10,7 +10,7 @@
 
 namespace static_type_info {
 
-  using TypeIndex = size_t;
+  using TypeIndex = uint64_t;
 
   template <class T> constexpr TypeIndex getTypeIndex() {
     return hash_64_fnv1a_const(getTypeName<T>());
